etls/despesas_governo/requirements.txt
pandas==1.4.3
boto3==1.24.33
selenium==4.3.0
beautifulsoup4==4.11.1
requests==2.28.1
psycopg2-binary==2.9.3